improvement (1.7.0更新)改进功能、一些小改进： 1.Blog支持按标签、按时间范围、关键字搜索查找。 2.Blog支持在临时信息流、日记列表中展示。 3.Blog列表改为按时间倒叙展示。 4.去除草稿箱功能。 5.检查更新的按钮移动到设置里了。 6.更换每日通知的提示方式。 7.编辑Blog退出时将会有提示。 8.为图库、视频库、保存日记、保存Blog添加Loading动画。 9.图库和视频库支持查看Blog中存储的图片、视频，并支持跳转。 10.更换纪念日的界面。

+++ Diary/app/src/main/java/cn/snowt/blog/BlogDetailActivity.java
@@ -39,6 +39,7 @@ import cn.snowt.diary.util.BaseUtils;
 import cn.snowt.diary.util.GlideEngine;
 import cn.snowt.diary.util.PermissionUtils;
 import cn.snowt.diary.util.SimpleResult;
+import cn.snowt.drawboard.DrawBoardActivity;
 import io.reactivex.Observable;
 import io.reactivex.ObservableEmitter;
 import io.reactivex.ObservableOnSubscribe;
@@ -236,12 +237,19 @@ public class BlogDetailActivity extends AppCompatActivity {
                 break;
             }
             case R.id.toolbar_del:{
-                SimpleResult result = blogService.deleteById(blogId);
-                if(result.getSuccess()){
-                    finish();
-                }else{
-                    BaseUtils.alertDialogToShow(context,"提示失败原因",result.getMsg());
-                }
+                androidx.appcompat.app.AlertDialog.Builder builder=new androidx.appcompat.app.AlertDialog.Builder(context);
+                builder.setTitle("提示：");
+                builder.setMessage("真的删除吗？");
+                builder.setNegativeButton("不删除，刚刚点错了", null);
+                builder.setPositiveButton("是的，我要删除",(dialog, which) -> {
+                    SimpleResult result = blogService.deleteById(blogId);
+                    if(result.getSuccess()){
+                        finish();
+                    }else{
+                        BaseUtils.alertDialogToShow(context,"提示失败原因",result.getMsg());
+                    }
+                });
+                builder.show();
                 break;
             }
             //编辑
